feat: Abstracted fetching tables. Also made a new class for database activities (documentation and how to use it will come tomorrow). Started with update and delete
