+++ app/src/main/java/no/usn/mob3000/data/network/SupabaseClientWrapper.kt
@@ -1,5 +1,6 @@
 package no.usn.mob3000.data.network
 
+
 import android.content.Context
 import io.github.jan.supabase.SupabaseClient
 import io.github.jan.supabase.createSupabaseClient
@@ -8,15 +9,7 @@ import io.github.jan.supabase.gotrue.Auth
 import io.github.jan.supabase.gotrue.auth
 import io.github.jan.supabase.postgrest.Postgrest
 import io.github.jan.supabase.realtime.Realtime
-import io.github.jan.supabase.postgrest.postgrest
-import kotlinx.serialization.KSerializer
-import kotlinx.serialization.Serializable
-import kotlinx.serialization.json.Json
-import kotlinx.serialization.encodeToString
-import kotlinx.serialization.json.JsonElement
-import kotlinx.serialization.json.JsonObject
 import no.usn.mob3000.MainActivity
-import kotlin.reflect.KClass
 import no.usn.mob3000.data.SecureEnvManager
 
 /**
@@ -84,43 +77,18 @@ object SupabaseClientWrapper {
         return supabaseClient
     }
 
+    /**
+     * Retrieves the current user's ID.
+     * @return The ID of the current user.
+     * @author Eirik
+     * @created 2024-10-28
+     */
     fun getCurrentUserId(): String? {
 
         val client = getClient()
         val userId = client.auth.currentSessionOrNull()?.user?.id ?: "user-id"
         return userId
     }
-
-    /**
-     * Inserts a new news article into the database.
-     *
-     *  TODO: Abstract this so it can be used for all other insert functions.
-     */
-    suspend fun <T> insertItem(table: String, item: T, serializer: KSerializer<T>): Result<Unit> {
-        return try {
-            val jsonElement: JsonElement = Json.encodeToJsonElement(serializer, item)
-                supabaseClient.postgrest[table].insert(jsonElement)
-                Result.success(Unit)
-        } catch (e: Exception) {
-            Result.failure(e)
-        }
-    }
-
-
-    /**
-     * Probably redundant as a similar data class exist elsewhere, here for testing
-     *
-     * TODO: Abstract :)
-     */
-    @Serializable
-    data class NewsItem(
-        val title: String,
-        val summary: String?,
-        val content: String?,
-        val is_published: Boolean,
-        val created_by: String
-    )
-
 
 
 }
